+++ nix/.stack.nix/cardano-wallet-core.nix
@@ -142,6 +142,7 @@ in { system, compiler, flags, pkgs, hsPkgs, pkgconfPkgs, ... }:
             (hsPkgs."quickcheck-state-machine" or (buildDepError "quickcheck-state-machine"))
             (hsPkgs."random" or (buildDepError "random"))
             (hsPkgs."retry" or (buildDepError "retry"))
+            (hsPkgs."safe" or (buildDepError "safe"))
             (hsPkgs."servant" or (buildDepError "servant"))
             (hsPkgs."servant-server" or (buildDepError "servant-server"))
             (hsPkgs."servant-swagger" or (buildDepError "servant-swagger"))
